[Refactor] 다른곳에 있던 서치 컨트롤러, 서비스, 레파지토리 독립
